+++ src/Entity/UserRoles.php
@@ -7,8 +7,14 @@ use Doctrine\ORM\Mapping as ORM;
  * UserRoles
  * @ORM\Entity(repositoryClass="FwsDoctrineAuth\Entity\Repository\UserRolesRepository")
  * @ORM\Table(name="user_roles", 
- * options={"collate"="latin1_swedish_ci", "charset"="latin1", "engine"="InnoDB"}, 
- * uniqueConstraints={@ORM\UniqueConstraint(name="role_idx", columns={"role"})}
+ *    options={
+ *        "collate"="latin1_swedish_ci", 
+ *        "charset"="latin1", 
+ *        "engine"="InnoDB"
+ *    }, 
+ *    uniqueConstraints={
+ *        @ORM\UniqueConstraint(name="role", columns={"role"})
+ *    }
  * )
  * @author Garry Childs <info@freedomwebservices.net>
  */