+++ src/Model/Service/ForgotPasswordModelFactory.php
@@ -5,6 +5,7 @@ namespace FwsDoctrineAuth\Model\Service;
 use Laminas\ServiceManager\Factory\FactoryInterface;
 use Interop\Container\ContainerInterface;
 use FwsDoctrineAuth\Model\ForgotPasswordModel;
+use FwsDoctrineAuth\Exception\DoctrineAuthException;
 use Doctrine\ORM\EntityManager;
 use Laminas\View\Renderer\PhpRenderer;