+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 
 Fixed bug in reset password model
 Removed dependency on jQuery
+
+0.1.14
+Added unique constraint to password reminder entity
+Added index annotations on entities
+Added use statement for DoctrineAuthException in ForgotPasswordModelFactory
+Set strict comparisons in ACL and Navigation Listener
+Added use statement for FwsDoctrineAuth\Entity\BaseUsers to AuthListener and NavigationListener